update default task for travis; refactoring and splitting of the indexer class into separate modules to separate concerns; start implementing find methods; update config file to allow development more easily; rip out old code to find files
