+++ .travis.yml
@@ -1,7 +1,7 @@
 sudo: false
 language: ruby
 cache: bundler
-script: bundle exec rake travis_spec
+script: bundle exec rake travis_ci 
 rvm:
   - 2.1.2
   - 2.2.4
