+++ app/controllers/docs_controller.rb
@@ -1,5 +1,6 @@
 class DocsController < ApplicationController
-  include IndexerHelper
+  include SolrMethods
+  include IndexerSetup
 
   # API call to get a full list of all solr documents modified between two times
   # @param [querystring] Parameters can be specified in the querystring
